% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mselection.R
\name{mselection}
\alias{mselection}
\title{mselection}
\usage{
mselection()
}
\value{
plot
}
\description{
this function studies the relationship between mammals which stay awake more than 12 hours and their order.
}
