% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualanalysis.R
\name{visualanalysis}
\alias{visualanalysis}
\title{visualanalysis}
\usage{
visualanalysis()
}
\value{
plot
}
\description{
this function is used to analyze data, in order to see visually the relationship between the way the animals feed themselves and the hours they spend awake.
}
